User friendly message for tempLocation. Closes #313
Temporary location is needed for BigQueryIO, in case of cloud based
runners at least one of `staging` or `temp` locations is required, for
location runners such requirement is not enforced thus it is possible
to end up with nasty NPE - see #313. Provide more user friendly
exception, with suggestion to set `temp` location.