CompressedSource: CompressedReader is never splittable
The only way it's safe to split a compressed file is if the file is not compressed. This can
only happen when the source itself is splittable, and that in turn will result in the inner
source's reader being returned. A CompressedReader will only be created in the event that
the file is NOT splittable. So remove all the logic handling splittable compressed readers,
and instead go with the logic when we know/assume the file is compressed.