Make classloader handling more compatible with JDK 9
From
http://www.oracle.com/technetwork/java/javase/9-relnote-issues-3704069.html:

The application class loader is no longer an instance of
java.net.URLClassLoader (an implementation detail that was never
specified in previous releases). Code that assumes that
ClassLoader::getSytemClassLoader returns a URLClassLoader object will
need to be updated. Note that Java SE and the JDK do not provide an API
for applications or libraries to dynamically augment the class path at
run-time.